LPS-64191 Extract baseline code into class
